+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/rules/IndentationRuleTest.kt
@@ -4820,11 +4820,7 @@ internal class IndentationRuleTest {
             indentationRuleAssertThat(code)
                 .addAdditionalRuleProvider { ParameterListWrappingRule() }
                 .withEditorConfigOverride(CODE_STYLE_PROPERTY to ktlint_official)
-                .hasLintViolationForAdditionalRule(
-                    2,
-                    20,
-                    "Parameter should be on a separate line (unless all parameters can fit a single line)",
-                )
+                .hasLintViolationForAdditionalRule(2, 20, "Parameter should start on a newline")
                 .hasLintViolations(
                     LintViolation(3, 1, "Unexpected indentation (19) (should be 12)"),
                     LintViolation(4, 1, "Unexpected indentation (19) (should be 12)"),